[collector] move lifecyclemanager and collector in internal
Moving lifecyclemanager and the collector components in their own internal packages. Will follow this up with some tests for each package.
